Move directional light logic to reactors
