+++ packages/engine/src/scene/systems/ShadowSystem.tsx
@@ -42,9 +42,7 @@ import config from '@etherealengine/common/src/config'
 import { isClient } from '@etherealengine/common/src/utils/getEnvironment'
 import { Engine, UUIDComponent } from '@etherealengine/ecs'
 import {
-  ComponentType,
   getComponent,
-  getMutableComponent,
   getOptionalComponent,
   hasComponent,
   removeComponent,
@@ -53,21 +51,12 @@ import {
   useOptionalComponent
 } from '@etherealengine/ecs/src/ComponentFunctions'
 import { ECSState } from '@etherealengine/ecs/src/ECSState'
-import { Entity, EntityUUID, UndefinedEntity } from '@etherealengine/ecs/src/Entity'
+import { Entity, UndefinedEntity } from '@etherealengine/ecs/src/Entity'
 import { createEntity, removeEntity, useEntityContext } from '@etherealengine/ecs/src/EntityFunctions'
 import { defineQuery, QueryReactor } from '@etherealengine/ecs/src/QueryFunctions'
 import { defineSystem, useExecute } from '@etherealengine/ecs/src/SystemFunctions'
 import { AnimationSystemGroup } from '@etherealengine/ecs/src/SystemGroups'
-import {
-  defineState,
-  dispatchAction,
-  getMutableState,
-  getState,
-  hookstate,
-  NO_PROXY,
-  State,
-  useHookstate
-} from '@etherealengine/hyperflux'
+import { defineState, getMutableState, getState, hookstate, NO_PROXY, useHookstate } from '@etherealengine/hyperflux'
 import { Vector3_Back } from '@etherealengine/spatial/src/common/constants/MathConstants'
 import {
   createPriorityQueue,
@@ -98,15 +87,11 @@ import { TransformComponent } from '@etherealengine/spatial/src/transform/compon
 import { XRLightProbeState } from '@etherealengine/spatial/src/xr/XRLightProbeSystem'
 import { isMobileXRHeadset } from '@etherealengine/spatial/src/xr/XRState'
 
-import { EngineState } from '@etherealengine/spatial/src/EngineState'
 import { useTexture } from '../../assets/functions/resourceLoaderHooks'
-import { GLTFNodeState, GLTFSnapshotAction } from '../../gltf/GLTFDocumentState'
-import { GLTFSnapshotState } from '../../gltf/GLTFState'
 import { DropShadowComponent } from '../components/DropShadowComponent'
 import { useMeshOrModel } from '../components/ModelComponent'
 import { RenderSettingsComponent } from '../components/RenderSettingsComponent'
 import { ShadowComponent } from '../components/ShadowComponent'
-import { SourceComponent } from '../components/SourceComponent'
 import { SceneObjectSystem } from './SceneObjectSystem'
 
 export const ShadowSystemState = defineState({
@@ -418,12 +403,7 @@ const updateDropShadowTransforms = () => {
   }
 }
 
-const directionalLightQuery = defineQuery([DirectionalLightComponent])
-const renderSettingsQuery = defineQuery([RenderSettingsComponent])
 const rendererQuery = defineQuery([RendererComponent])
-
-// Store UUIDs in a map because getComponent isn't available on exit query
-const directionalLightMap = new Map<Entity, EntityUUID>()
 
 const execute = () => {
   if (!isClient) return
@@ -432,41 +412,6 @@ const execute = () => {
   if (!useShadows) {
     updateDropShadowTransforms()
     return
-  }
-
-  for (const entity of directionalLightQuery.enter()) {
-    const uuid = getComponent(entity, UUIDComponent)
-    directionalLightMap.set(entity, uuid)
-  }
-
-  for (const entity of directionalLightQuery.exit()) {
-    const uuid = directionalLightMap.get(entity)
-    if (!uuid) continue
-
-    // CSM directional lights should be their own component
-    for (const renderSettingsEntity of renderSettingsQuery()) {
-      const renderSettings = getComponent(renderSettingsEntity, RenderSettingsComponent)
-      if (renderSettings.primaryLight === uuid) {
-        getMutableComponent(renderSettingsEntity, RenderSettingsComponent).merge({
-          csm: false,
-          primaryLight: '' as EntityUUID
-        })
-        if (getState(EngineState).isEditing) {
-          const source = getComponent(renderSettingsEntity, SourceComponent)
-          const node = GLTFNodeState.getMutableNode(renderSettingsEntity)
-          const renderSettingsExt = node.extensions[RenderSettingsComponent.jsonID] as State<
-            ComponentType<typeof RenderSettingsComponent>
-          >
-          renderSettingsExt.merge({
-            csm: false,
-            primaryLight: '' as EntityUUID
-          })
-          const snapshot = GLTFSnapshotState.cloneCurrentSnapshot(source)
-          dispatchAction(GLTFSnapshotAction.createSnapshot(snapshot))
-        }
-      }
-    }
-    directionalLightMap.delete(entity)
   }
 
   for (const entity of rendererQuery()) {
